requirements.txt
Flask==2.3.3
# If you want to use gunicorn in production (optional):
# gunicorn==20.1.0
pytest==7.4.0
requests==2.31.0
azure-cosmos
werkzeug
azure-storage-blob
pytest-mock